v8: bundlify javascript code
__openerp__.py 'js' dictionary doesn't work since before v8.0 final
release. We need to add it to the web bundle assets for the JS to
even be seen by the browser.
The upstream bundling was done in odoo commit 1c4e2cfd22f2 for the
included modules, and the 'js' dictionary support broken with that
or in an earlier/later related change.