Populate system font registry before using it in PDF, fully fixing glyph coverage
In Odoo v8 the font population seems to have reworked a bit, so this adapts to it.

We now need to call font_scan on the company, which core reportlab support does
as well since v8. Without this, the system fonts aren't actually used, because
the listing in customfonts module isn't filled yet. If it is filled already,
then thanks to lazy=True, it isn't rescanned, as something has done it already.
If fonts are added to the system, manual font scanning can be called via
company form "Report Configuration" tab "(reload fonts)" link, which calls
font_scan with lazy=False.
This change ensures that fonts work properly when the first reportlab usage
is via this module, without having to manually scan them or trigger it from
elsewhere. wkhtmltopdf doesn't use this, so it's actually unlikely anything
else triggers it.

Ideally this module would be ported to QWeb/wkhtmltopdf as well, or at least
various other potential changes that have happened to the core trml2pdf.py and
other file copies brought over if appropriate.